// template/react-starter-kit/src/.generated/api/model/userFilterStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Nest Starter Kit
 * NestJS + MikroORM Starter Kit API
 * OpenAPI spec version: 1.0.0
 */

export type UserFilterStatus = typeof UserFilterStatus[keyof typeof UserFilterStatus];


export const UserFilterStatus = {
  active: 'active',
  banned: 'banned',
  deleted: 'deleted',
} as const;