// template/react-starter-kit/src/.generated/api/model/usersControllerGetUsersSortItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Nest Starter Kit
 * NestJS + MikroORM Starter Kit API
 * OpenAPI spec version: 1.0.0
 */

export type UsersControllerGetUsersSortItem = typeof UsersControllerGetUsersSortItem[keyof typeof UsersControllerGetUsersSortItem];


export const UsersControllerGetUsersSortItem = {
  name: 'name',
  email: 'email',
  role: 'role',
  twoFactorEnabled: 'twoFactorEnabled',
  createdAt: 'createdAt',
  updatedAt: 'updatedAt',
  id: 'id',
} as const;